Explicit cast from int64 to uint32.

+++ src/parser/dbclineparser.cpp
@@ -151,7 +151,7 @@ uint32_t DbcLineParser::ParseMultiTrans(std::vector<std::string>& outnodes, std:
 
   if (chunks.size() >= 3 && chunks[0] == "BO_TX_BU_")
   {
-    ret = clear_msgid(atoll(chunks[1].c_str()));
+    ret = clear_msgid(static_cast<uint32_t>(atoll(chunks[1].c_str())));
 
     if (ret != 0)
     {
@@ -453,7 +453,7 @@ bool DbcLineParser::ParseCommentLine(Comment_t* cm, const std::string& line)
           // 1 CM_ marker
           // 2 target (message or signal)
           // 3 msg id
-          uint32_t id = static_cast<uint32_t>((atoll(meta[2].c_str())));
+          uint32_t id = static_cast<uint32_t>(atoll(meta[2].c_str()));
 
           // clear message id from high 3 bits
           cm->MsgId = clear_msgid(id);
@@ -510,7 +510,7 @@ bool DbcLineParser::ParseAttributeLine(AttributeDescriptor_t* attr, const std::s
       if (items.size() > 4 && items[1] == "GenMsgCycleTime" && items[2] == "BO_")
       {
         attr->Type = AttributeType::CycleTime;
-        attr->MsgId = clear_msgid(atoll(items[3].c_str()));
+        attr->MsgId = clear_msgid(static_cast<uint32_t>(atoll(items[3].c_str())));
         // read value of ms of cycle time for the current message
         attr->Value = atoi(items[4].c_str());
         ret = true;
@@ -549,7 +549,7 @@ bool DbcLineParser::ParseValTableLine(Comment_t* comm, const std::string& line)
 
       if ((items.size() >= 3) && (items.back() == ";") && (items.size() % 2 == 0))
       {
-        comm->MsgId = (clear_msgid(atoll(items[1].c_str())));
+        comm->MsgId = (clear_msgid(static_cast<uint32_t>(atoll(items[1].c_str()))));
         comm->SigName = items[2];
         comm->Text = "";
         comm->ca_target = CommentTarget::Signal;
